+++ tests/components/api.test.js
@@ -676,12 +676,7 @@ describe('API tests', () => {
           expectHandled(SUPPORTABILITY_METRIC_CHANNEL, ['API/wrapLogger/called'])
 
           expectEmitted('wrap-logger-start', [expect.any(Array), expect.any(Object), 'myObservedLogger'])
-          expectEmitted('wrap-logger-end', [['test1'], expect.any(Object), undefined, expect.objectContaining({
-            duration: expect.any(Number),
-            isLongTask: false,
-            methodName: 'myObservedLogger',
-            thrownError: undefined
-          })])
+          expectEmitted('wrap-logger-end', [['test1'], expect.any(Object), undefined])
 
           expectHandled(SUPPORTABILITY_METRIC_CHANNEL, ['API/logging/info/called'])
           expectHandled('log', [expect.any(Number), 'test1', {}, 'INFO', undefined])
@@ -709,12 +704,7 @@ describe('API tests', () => {
           expectHandled(SUPPORTABILITY_METRIC_CHANNEL, ['API/wrapLogger/called'])
 
           expectEmitted('wrap-logger-start', [expect.any(Array), expect.any(Object), randomMethodName])
-          expectEmitted('wrap-logger-end', [['test1'], expect.any(Object), undefined, expect.objectContaining({
-            duration: expect.any(Number),
-            isLongTask: false,
-            methodName: randomMethodName,
-            thrownError: undefined
-          })])
+          expectEmitted('wrap-logger-end', [['test1'], expect.any(Object), undefined])
 
           expectHandled(SUPPORTABILITY_METRIC_CHANNEL, ['API/logging/warn/called'])
           expectHandled('log', [expect.any(Number), 'test1', {}, 'warn', undefined])
@@ -735,12 +725,7 @@ describe('API tests', () => {
           expectHandled(SUPPORTABILITY_METRIC_CHANNEL, ['API/wrapLogger/called'])
 
           expectEmitted('wrap-logger-start', [expect.any(Array), expect.any(Object), randomMethodName])
-          expectEmitted('wrap-logger-end', [['test1', { test2: 2 }, ['test3'], true, 1], expect.any(Object), undefined, expect.objectContaining({
-            duration: expect.any(Number),
-            isLongTask: false,
-            methodName: randomMethodName,
-            thrownError: undefined
-          })])
+          expectEmitted('wrap-logger-end', [['test1', { test2: 2 }, ['test3'], true, 1], expect.any(Object), undefined])
         })
 
         test('wrapped function should still behave as intended', () => {