+++ tests/specs/third-party-compatibility/run-test.js
@@ -74,5 +74,4 @@ export default async function runTest ({
       ajaxMetricsHarvests
     }
   )
-  await browser.collectCoverage()
 }